- allow RewriteFactor[*], RewriteFactor[+] in xgrs of backend, too; emit them from frontend; - no hacky special long numbers for iteration ranges and constants anymore, the default int is now used consistently - same literal syntax now everwhere:   - in sequences: allow hex numbers; allow set/map initialization (at variable assignment)   - in shell: allow float, double, enum at attribute initialization; allow hex numbers; allow set/map literal initialization (at variable assignment)   - in grs importer: allow hex numbers; allow float, double, enum at unquoted attribute initialization   - in grs exporter: write unquoted attribute initialization, in same literal format as used everywhere else - accept .grsi suffix in porter for .grs files

+++ frontend/de/unika/ipd/grgen/ast/RangeSpecNode.java
@@ -25,8 +25,8 @@ public class RangeSpecNode extends BaseNode {
 		setName(RangeSpecNode.class, "range spec");
 	}
 
-	/** Constant, signaling if one bound is bounded. */
-	public static final long UNBOUND = Long.MAX_VALUE;
+	/** Constant, signaling if upper bound is bounded. */
+	public static final long UNBOUND = Integer.MAX_VALUE;
 
 	/** The upper and lower bound. */
 	private long lower, upper;
